+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `sendNotificationEmailOptionDefaultUnchecked` to `FormApprovalConfiguration`
+
 ## [11.0.7] - 2025-07-10
 
 ### Added